refactor: LostItemId 도입
